// utils/interface.ts
export interface HeroData {
  title: string;
  heroDescription: string;
  heroImgDesktop: string;
  heroImgMobile: string;
  buttonLink: string;
}

export interface NewsColumnData {
  id: string;
  title: string;
  description: string;
  link: string;
}

export interface NewsRowData {
  id: string;
  title: string;
  description: string;
  link: string;
  imgSrc: string;
}